Recombine files ProcessTests.cs and ProcessTests.netcoreapp.cs that aren't supported anymore
